Expose unit test results to GitLab UI
GitLab can expose unit test results in the pipeline UI,
but only if the results are in JUnit format, uploaded as
an artifact.

This commit converts those results to JUnit format, by
running saxonb-xslt on an NUnit->JUnit XSLT transform.

The transform is from the project nunit/nunit-transforms, published
under the MIT license by Paul Hicks. The license for the transform
we use here can also be found in the original project repository:

https://github.com/nunit/nunit-transforms/tree/master/nunit3-junit